requirements.txt
asgiref==3.8.1
Django==5.1.5
django-htmx==1.22.0
et_xmlfile==2.0.0
greenlet==3.1.1
lxml==5.3.1
numpy==2.2.2
openpyxl==3.1.5
pandas==2.2.3
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
six==1.17.0
SQLAlchemy==2.0.37
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
xlwings==0.33.7
